requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.6
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.7
dulwich==0.21.5
fastapi==0.101.1
filelock==3.12.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jsonschema==4.17.3
keyring==24.2.0
magic-filter==1.0.11
more-itertools==10.1.0
msgpack==1.0.5
multidict==6.0.4
packaging==23.1
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pydantic==2.1.1
pydantic_core==2.4.0
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
pywin32-ctypes==0.2.2
qrcode==7.4.2
rapidfuzz==2.15.1
redis==5.0.0
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.3
yarl==1.9.2
zipp==3.16.2
